Start Rewrite
This project is a complete mess. It currently doesnt compile and it has
a butt ton of problems.

I have an idea for an algorithm that will make this whole thing simpler.

basically, it splits it up into 2 phases. spying phase(info gathering)
and fashion phase(formatted printing).

spying phase has an array, it does dir walking to gather all the info.
if its a directory, add another element to the entries array and add the
info there. if its a file, add the info to its parent struct. basically
generating a tree of all the entries to be printed, their type, and a
bunch of other info.

the fashion phase prints the given info by basically traversing the lil
entry tree. Theoretically, the printing module can be swapped out so we
can support multiple printing schemes like json or something.