Add: parameter titling of output behavior
this adds the parameter little title like `<parameter>:` then new line.
The idea is to mimic `ls`s behavior when you give multiple parameters.
If more than 1 is given, the titles show up. Anyway I mimiced this
behavior.

See cat doesnt have this behavior to my knowledge so I will add a flag
to disable this. I personally am fine with it as default and I am the
main audience.