WIP: ls implementation
The idea is to have 2 stage. a stage based around efficient and modular
information gathering. Called the `spy` phase. the other stage is around
modular easy to program printing.

The spy phase has entries which hold the directory entries data, like
stat and name and stuff. The spy entries go through processing in the
states with the printing entries. the printing entries are structs that
hold all the information to be printed out.

The reason why its 2 phase is we have to do preprocessing of the data.
like correct spacing to make a nice looking table. Anyway the best way
to do this is by gathering the information first and then printing. this
was the solution that I came up with to also have it nice and modular.