Add: calculate feature & name state implementation
calculate feature currently is to calculate the formatting length. This
will probably change cause its a lil wonky rn but this is an implementation.

Anyway also do an implementation of it in `name`. This works fine as
expected. When printing out the entry structs because of another bug, I
found a use after free of stack data. Fixing that was just allocating on
the heap.