Add: size field for `-l`, Add: proper column space
size field was missing in the `-l` implementation so I added it. This
was trivial to add because of the architecture. It makes it so I have to
implement 3 functions and update 3 switch statements. I _could_ do some
other shenanigans if I were to use `else` in the switch statements.

proper column spacing was much needed. I was literally just printing out
the max size before and it was going *ok*. This is fully implemented as
of right now.

in the future, itll probably be best to refactor into its own `column`
logic where if each only has a single state (other than `end`) then do
that (kinda like `ls`) else do full multi state per entry
table. *anyway* its fine enough for now so lets not touch it and get
this done enough.